refactor: move domain usecases to its own entities
