+++ src/data/protocols/db/account/add-account-repository.ts
@@ -1,4 +1,4 @@
-import { AddAccountModel } from '@src/domain/usecases/add-account'
+import { AddAccountModel } from '@src/domain/usecases/account/add-account'
 import { AccountModel } from '@src/domain/models/account'
 
 export interface AddAccountRepository {
