+++ src/data/usecases/add-account/db-add-account-usecase-protocols.ts
@@ -1,4 +1,4 @@
 export * from '@src/data/protocols/crypto/hasher'
 export * from '@src/data/protocols/db/account/add-account-repository'
 export * from '@src/domain/models/account'
-export * from '@src/domain/usecases/add-account'
+export * from '@src/domain/usecases/account/add-account'
