+++ src/data/usecases/load-account-by-token/db-load-account-by-token-protocols.ts
@@ -1,4 +1,4 @@
-export * from '@src/domain/usecases/load-account-by-token'
+export * from '@src/domain/usecases/account'
 export { Decrypter } from '@src/data/protocols/crypto'
 export { AccountModel } from '@src/domain/models/account'
 export { LoadAccountByTokenRepository } from '@src/data/protocols/db'
