+++ src/domain/usecases/account/authentication.ts
@@ -1,4 +1,4 @@
-import { AccountModel } from '../models/account'
+import { AccountModel } from '../../models/account'
 
 export type AuthModel = Omit<AccountModel, 'name' | 'id'>
