+++ src/domain/usecases/survey/add-survey.ts
@@ -1,4 +1,4 @@
-import { SurveyModel } from '../models/survey'
+import { SurveyModel } from '../../models/survey'
 
 export interface AddSurveyUseCase<T> {
   add(data: SurveyModel): Promise<T>
