+++ src/domain/usecases/survey/index.ts
@@ -1,1 +1,2 @@
 export * from './load'
+export * from './add-survey'
