+++ src/domain/usecases/survey/load.ts
@@ -1,0 +1,5 @@
+import { SurveyModel } from '@src/domain/models/survey'
+
+export interface LoadSurveyUseCase {
+  load(): Promise<SurveyModel[]>
+}
