+++ src/main/factories/usecases/add-survey/db-add-survey-usecase-factory.ts
@@ -1,6 +1,6 @@
 import { DbAddSurveyUseCase } from '@src/data/usecases/add-survey/db-add-survey-usecase'
 import { DbAddSurveyResult } from '@src/data/usecases/add-survey/db-add-survey-usecase-result'
-import { AddSurveyUseCase } from '@src/domain/usecases/add-survey'
+import { AddSurveyUseCase } from '@src/domain/usecases/survey/add-survey'
 import { SurveyMongoRepository } from '@src/infra/db/mongodb/survey/survey-mongo-repository'
 
 export const makeDbAddSurveyUseCase =
