+++ src/main/factories/usecases/authentication/db-authentication-usecase-factory.ts
@@ -1,6 +1,6 @@
 import { DbAddAccountUseCase } from '@src/data/usecases/add-account/db-add-account-usecase'
 import { DbAddAccountResult } from '@src/data/usecases/add-account/db-add-account-usecase-result'
-import { AddAccountUseCase } from '@src/domain/usecases/add-account'
+import { AddAccountUseCase } from '@src/domain/usecases/account/add-account'
 import { BcryptAdapter } from '@src/infra/crypto/bcrypt/bcrypt-adapter'
 import { AccountMongoRepository } from '@src/infra/db/mongodb'
 import { ENV } from '@src/main/config/env'
