+++ src/main/factories/usecases/load-account-by-token/db-load-account-by-token-factory.ts
@@ -1,6 +1,6 @@
 import { DbLoadAccountByTokenUsecase } from '@src/data/usecases/load-account-by-token/db-load-account-by-token'
 import { DbLoadAccountByTokenUsecaseResult } from '@src/data/usecases/load-account-by-token/db-load-account-by-token-result'
-import { LoadAccountByTokenUseCase } from '@src/domain/usecases/load-account-by-token'
+import { LoadAccountByTokenUseCase } from '@src/domain/usecases/account/load-account-by-token'
 import { TokenGeneratorAdapter } from '@src/infra/crypto/jwt/jwt-adpter'
 import { AccountMongoRepository } from '@src/infra/db/mongodb'
 import { ENV } from '@src/main/config/env'
