+++ src/presentation/controllers/account/login/login-controller-protocols.ts
@@ -1,2 +1,2 @@
 export * from '@src/presentation/protocols'
-export * from '@src/domain/usecases/authentication'
+export * from '@src/domain/usecases/account'
