+++ src/presentation/controllers/account/signup/signup-controller-protocols.ts
@@ -1,4 +1,3 @@
 export * from '@src/presentation/protocols'
-export * from '@src/domain/usecases/add-account'
-export * from '@src/domain/usecases/authentication'
+export * from '@src/domain/usecases/account'
 export * from '@src/domain/models/account'
