+++ src/presentation/controllers/survey/add/add-survey-protocols.ts
@@ -1,3 +1,3 @@
 export * from '@src/presentation/protocols'
-export * from '@src/domain/usecases/add-survey'
+export * from '@src/domain/usecases/survey'
 export * from '@src/data/usecases/add-survey/db-add-survey-usecase-result'
