+++ src/presentation/middlewares/auth-middleware-protocols.ts
@@ -1,3 +1,3 @@
 export * from '@src/presentation/protocols'
-export * from '@src/domain/usecases/load-account-by-token'
+export * from '@src/domain/usecases/account'
 export { DbLoadAccountByTokenUsecaseResult } from '@src/data/usecases/load-account-by-token/db-load-account-by-token-result'
